crete SearchQuery builder
