+++ packages/server/src/modules/entities/entities.search.test.ts
@@ -11,7 +11,6 @@ import "ts-jest";
 import { supertestConfig } from "..";
 import app from "../../Server";
 import { prepareStatement } from "@models/statement/statement.test";
-import { filterEntitiesByWildcard } from "@models/entity/response_search";
 
 describe("Entities search (requests)", function () {
   describe("empty data", () => {
@@ -285,83 +284,6 @@ describe("Entities search (params)", function () {
           done();
         });
       });
-    });
-  });
-});
-
-describe("Entities search - advanced label search", function () {
-  let db: Db;
-  const rand = Math.random().toString();
-
-  const [, nameEntity] = prepareEntity();
-  nameEntity.label = "Evelín Teměř Jr.";
-  nameEntity.id = `${nameEntity.label}-${nameEntity.id}`;
-  nameEntity.class = EntityClass.Person;
-
-  const [, eventEntity] = prepareEntity();
-  eventEntity.label = "TRP yyyy-mm-dd: during";
-  eventEntity.id = `${eventEntity.label}-${eventEntity.id}`;
-  eventEntity.class = EntityClass.Event;
-
-  beforeAll(async () => {
-    db = new Db();
-    await db.initDb();
-
-    await nameEntity.save(db.connection);
-    await eventEntity.save(db.connection);
-  });
-
-  afterAll(async () => {
-    await deleteEntities(db);
-    await db.close();
-  });
-
-  describe("search by word", () => {
-    it("should return found a result when searching word by word", async () => {
-      const words = eventEntity.label
-        .replace(/[^a-zA-Z0-9]+/g, " ")
-        .split(" ")
-        .map((w) => w.trim());
-      for (const word of words) {
-        const entities = await filterEntitiesByWildcard(
-          db.connection,
-          false,
-          undefined,
-          `*${word}*`,
-          undefined
-        );
-        expect(entities).toHaveLength(1);
-      }
-    });
-  });
-
-  describe("search by parts", () => {
-    it("should return found a result when searching by parts", async () => {
-      const parts = [
-        "Evelín Teměř Jr",
-        "Evelín Teměř Jr.",
-        "Teměř",
-        "Evelín",
-        "Jr",
-        "Jr.",
-        "Temě*",
-        "Evel*",
-        "Evelín Jr",
-      ];
-      for (const part of parts) {
-        const entities = await filterEntitiesByWildcard(
-          db.connection,
-          false,
-          undefined,
-          part,
-          undefined
-        );
-        try {
-          expect(entities).toHaveLength(1);
-        } catch (e) {
-          throw new Error(`${part} not sattisfied`);
-        }
-      }
     });
   });
 });
